+++ tests/test_ecs.py
@@ -239,6 +239,7 @@ def test_tag_query() -> None:
     world = tcod.ecs.World()
     assert not set(world.Q.all_of(tags=["A"]))
     world["A"].tags.add("A")
+    assert set(world.Q.all_of(tags=["A"])) == {world["A"]}
     world["A"].tags.add("A")  # Cover redundant add
     assert set(world.Q.all_of(tags=["A"])) == {world["A"]}
     world["A"].tags.remove("A")